+++ game/src/GDataBuffer.cpp
@@ -149,8 +149,8 @@ GDataBuffer::~GDataBuffer()
 
 bool GDataBuffer::changeCount(guint offset,guint stride,guint count)
 {
-    m_Offset=GX_CAST_S(guint, offset);
-    m_Stride=GX_CAST_S(guint, stride);
+    m_Offset=GX_CAST_S(guint16, offset);
+    m_Stride=GX_CAST_S(guint16, stride);
     
     return changeBytes(count*m_Stride);
 }
